+++ mycodo/scripts/mycodo_wrapper.c
@@ -158,6 +158,18 @@ int main(int argc, char *argv[]) {
 			strncat(updateScript, "/upgrade_commands.sh enable-pigpiod-high", sizeof(updateScript));
 			system(updateScript);
 
+		} else if (strcmp(argv[1], "enable_pigpiod_disabled") == 0) {
+
+            char path[255];
+            strncpy(path, argv[0], sizeof(path));
+            dirname(path);
+
+			char updateScript[255];
+			strncpy(updateScript, "/bin/bash ", sizeof(updateScript));
+			strncat(updateScript, path, sizeof(updateScript));
+			strncat(updateScript, "/upgrade_commands.sh enable-pigpiod-disabled", sizeof(updateScript));
+			system(updateScript);
+
 		}
 	} else {